requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
playwright==1.37.0
pandas==2.2.2
pdfplumber==0.8.0
python-magic==0.4.27
python-dotenv==1.0.0
PyPDF2==3.0.0